pyproject.toml
[tool.poetry]
name = "cd4ml"
version = "0.1.0"
description = "cd4ml on github"
authors = ["Claudio Davi <cdavisouza@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
absl-py = "0.11.0"
appdirs = "1.4.4"
astunparse = "1.6.3"
atpublic = "2.1.2"
cachetools = "4.2.0"
certifi = "2020.12.5"
cffi = "1.14.4"
chardet = "4.0.0"
colorama = "0.4.4"
commonmark = "0.9.1"
configobj = "5.0.6"
decorator = "4.4.2"
dictdiffer = "0.8.1"
distro = "1.5.0"
dpath = "2.0.1"
dulwich = "0.20.15"
dvc = {extras = ["gcs", "aws"], version = "^1.11.13"}
flatbuffers = "1.12"
flatten-dict = "0.3.0"
"flufl.lock" = "3.2"
ftfy = "5.8"
funcy = "1.15"
future = "0.18.2"
gast = "0.3.3"
gitdb = "4.0.5"
GitPython = "3.1.12"
google-api-core = "1.24.1"
google-auth = "1.24.0"
google-auth-oauthlib = "0.4.2"
google-cloud-core = "1.5.0"
google-cloud-storage = "1.19.0"
google-crc32c = "1.1.0"
google-pasta = "0.2.0"
google-resumable-media = "1.2.0"
googleapis-common-protos = "1.52.0"
grandalf = "0.6"
grpcio = "1.32.0"
h5py = "2.10.0"
idna = "2.10"
jsonpath-ng = "1.5.2"
Keras-Preprocessing = "1.1.2"
mailchecker = "4.0.3"
Markdown = "3.3.3"
nanotime = "0.5.2"
networkx = "2.5"
numpy = "1.19.5"
oauthlib = "3.1.0"
opt-einsum = "3.3.0"
packaging = "20.8"
pathlib2 = "2.3.5"
pathspec = "0.8.1"
phonenumbers = "8.12.16"
ply = "3.11"
protobuf = "3.14.0"
pyasn1 = "0.4.8"
pyasn1-modules = "0.2.8"
pycparser = "2.20"
pydot = "1.4.1"
Pygments = "2.7.4"
pygtrie = "2.3.2"
pyparsing = "2.4.7"
python-benedict = "0.22.4"
python-dateutil = "2.8.1"
python-slugify = "4.0.1"
pytz = "2020.5"
PyYAML = "5.3.1"
requests = "2.25.1"
requests-oauthlib = "1.3.0"
rich = "9.8.0"
rsa = "4.7"
"ruamel.yaml" = "0.16.12"
"ruamel.yaml.clib" = "0.2.2"
shortuuid = "1.0.1"
shtab = "1.3.4"
six = "1.15.0"
smmap = "3.0.4"
tabulate = "0.8.7"
tensorboard = "2.4.0"
tensorboard-plugin-wit = "1.7.0"
tensorflow = "2.4.0"
tensorflow-estimator = "2.4.0"
termcolor = "1.1.0"
text-unidecode = "1.3"
toml = "0.10.2"
tqdm = "4.56.0"
typing-extensions = "3.7.4.3"
urllib3 = "1.26.2"
voluptuous = "0.12.1"
wcwidth = "0.2.5"
Werkzeug = "1.0.1"
wrapt = "1.12.1"
xmltodict = "0.12.0"
"zc.lockfile" = "2.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
